attkey新形式に対応

+++ Common/ashx/Common.ashx
@@ -26,28 +26,34 @@ Public Class CommonAjax : Implements IHttpHandler
         Dim sJSON As String = ""
 
         Try
-            sAttKey = cCki.Get_Cookies("AttKey")
-            cDB.AddWithValue("@table_schema", "lineojtdb")
-            Dim DelTableArray As New List(Of String)
+            Dim sUser_ID As String = cCki.Get_Cookies("User_ID")
+            If cCki.Get_Cookies(sUser_ID & "AttKey") <> "" Then
+                sAttKey = cCki.Get_Cookies(sUser_ID & "AttKey")
+                cDB.AddWithValue("@table_schema", "lineojtdb")
+                Dim DelTableArray As New List(Of String)
 
-            sSQL.Clear()
-            sSQL.Append(" SELECT")
-            sSQL.Append(" TABLE_NAME")
-            sSQL.Append(" FROM")
-            sSQL.Append(" information_schema.tables")
-            sSQL.Append(" WHERE TABLE_SCHEMA = @table_schema")
-            sSQL.Append("   AND TABLE_NAME LIKE '%" & sAttKey & "%'")
-            cDB.SelectSQL(sSQL.ToString)
-            Do Until Not cDB.ReadDr
-                DelTableArray.Add(cDB.DRData("TABLE_NAME"))
-            Loop
+                sSQL.Clear()
+                sSQL.Append(" SELECT")
+                sSQL.Append(" TABLE_NAME")
+                sSQL.Append(" FROM")
+                sSQL.Append(" information_schema.tables")
+                sSQL.Append(" WHERE TABLE_SCHEMA = @table_schema")
+                sSQL.Append("   AND TABLE_NAME LIKE '%" & sAttKey & "%'")
+                cDB.SelectSQL(sSQL.ToString)
+                Do Until Not cDB.ReadDr
+                    DelTableArray.Add(cDB.DRData("TABLE_NAME"))
+                Loop
 
-            sSQL.Clear()
-            sSQL.Append(" DROP TABLE")
-            sSQL.Append(" IF Exists")
-            sSQL.Append(" " & String.Join(",", DelTableArray))
-            cDB.ExecuteSQL(sSQL.ToString)
-
+                If DelTableArray.Count > 0 Then
+                    sSQL.Clear()
+                    sSQL.Append(" DROP TABLE")
+                    sSQL.Append(" IF Exists")
+                    sSQL.Append(" " & String.Join(",", DelTableArray))
+                    cDB.ExecuteSQL(sSQL.ToString)
+                End If
+            Else
+                sStatus = "PASS"
+            End If
         Catch ex As Exception
             sRet = ex.Message
         Finally
